package/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.0
idna==2.8
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
